wrap/unwrap http client
